Move OcrData to common package
