+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/model/common/OcrData.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanprocessor.entity;
+package uk.gov.hmcts.reform.bulkscanprocessor.model.common;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -9,7 +9,7 @@ public class OcrData {
     @JsonProperty("Metadata_file")
     public final List<OcrDataField> fields;
 
-    public OcrData(List<OcrDataField> fields) {
+    public OcrData(@JsonProperty("Metadata_file") List<OcrDataField> fields) {
         this.fields = fields;
     }
 }
